OK-509: Järjestelty valintakoekutsujen koodia paremmin
