+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintakoekutsut/valintakoekutsut-table.tsx
@@ -5,10 +5,10 @@ import ListTable, {
   ListTableColumn,
 } from '@/app/components/table/list-table';
 import { useTranslations } from '@/app/hooks/useTranslations';
-import { ValintakoeKutsuItem } from '@/app/hooks/useValintakoekutsut';
 import { Box } from '@mui/material';
 import { useMemo, useState } from 'react';
 import { toFormattedDateTimeString } from '@/app/lib/localization/translation-utils';
+import { ValintakoeKutsuItem } from '@/app/lib/types/valintakoekutsut-types';
 
 const TRANSLATIONS_PREFIX = 'valintakoekutsut.taulukko';
