+++ src/app/lib/valintalaskentakoostepalvelu.test.ts
@@ -76,32 +76,32 @@ describe('getValintakoekutsutData', () => {
       ),
       hakemuksetByOid: {
         '1.2.246.562.11.00000000000001796027': {
-          oid: '1.2.246.562.11.00000000000001796027',
-          henkiloOid: '1.2.246.562.24.69259807406',
+          hakemusOid: '1.2.246.562.11.00000000000001796027',
+          hakijaOid: '1.2.246.562.24.69259807406',
           etunimet: 'Ruhtinas',
           sukunimi: 'Nukettaja',
           hakijanNimi: 'Nukettaja Ruhtinas',
           asiointikieliKoodi: 'fi',
         },
         '1.2.246.562.11.00000000000001793706': {
-          oid: '1.2.246.562.11.00000000000001793706',
-          henkiloOid: '1.2.246.562.24.25732574711',
+          hakemusOid: '1.2.246.562.11.00000000000001793706',
+          hakijaOid: '1.2.246.562.24.25732574711',
           etunimet: 'Kreivi',
           sukunimi: 'Dacula',
           hakijanNimi: 'Dacula Kreivi',
           asiointikieliKoodi: 'sv',
         },
         '1.2.246.562.11.00000000000001790371': {
-          oid: '1.2.246.562.11.00000000000001790371',
-          henkiloOid: '1.2.246.562.24.14598775927',
+          hakemusOid: '1.2.246.562.11.00000000000001790371',
+          hakijaOid: '1.2.246.562.24.14598775927',
           etunimet: 'Puru',
           sukunimi: 'Purukumi',
           hakijanNimi: 'Purukumi Puru',
           asiointikieliKoodi: 'fi',
         },
         '1.2.246.562.11.00000000000001543832': {
-          oid: '1.2.246.562.11.00000000000001543832',
-          henkiloOid: '1.2.246.562.24.30476885816',
+          hakemusOid: '1.2.246.562.11.00000000000001543832',
+          hakijaOid: '1.2.246.562.24.30476885816',
           etunimet: 'Haamu',
           sukunimi: 'Hui',
           hakijanNimi: 'Hui Haamu',
@@ -166,32 +166,32 @@ describe('getValintakoekutsutData', () => {
       ),
       hakemuksetByOid: {
         '1.2.246.562.11.00000000000001796027': {
-          oid: '1.2.246.562.11.00000000000001796027',
-          henkiloOid: '1.2.246.562.24.69259807406',
+          hakemusOid: '1.2.246.562.11.00000000000001796027',
+          hakijaOid: '1.2.246.562.24.69259807406',
           etunimet: 'Ruhtinas',
           sukunimi: 'Nukettaja',
           hakijanNimi: 'Nukettaja Ruhtinas',
           asiointikieliKoodi: 'fi',
         },
         '1.2.246.562.11.00000000000001793706': {
-          oid: '1.2.246.562.11.00000000000001793706',
-          henkiloOid: '1.2.246.562.24.25732574711',
+          hakemusOid: '1.2.246.562.11.00000000000001793706',
+          hakijaOid: '1.2.246.562.24.25732574711',
           etunimet: 'Kreivi',
           sukunimi: 'Dacula',
           hakijanNimi: 'Dacula Kreivi',
           asiointikieliKoodi: 'sv',
         },
         '1.2.246.562.11.00000000000001790371': {
-          oid: '1.2.246.562.11.00000000000001790371',
-          henkiloOid: '1.2.246.562.24.14598775927',
+          hakemusOid: '1.2.246.562.11.00000000000001790371',
+          hakijaOid: '1.2.246.562.24.14598775927',
           etunimet: 'Puru',
           sukunimi: 'Purukumi',
           hakijanNimi: 'Purukumi Puru',
           asiointikieliKoodi: 'fi',
         },
         '1.2.246.562.11.00000000000001543832': {
-          oid: '1.2.246.562.11.00000000000001543832',
-          henkiloOid: '1.2.246.562.24.30476885816',
+          hakemusOid: '1.2.246.562.11.00000000000001543832',
+          hakijaOid: '1.2.246.562.24.30476885816',
           etunimet: 'Haamu',
           sukunimi: 'Hui',
           hakijanNimi: 'Hui Haamu',
